Converts all `uint8_t` and `uint16_t` casts to the functional style.

+++ Processors/Z80/Implementation/Z80Base.cpp
@@ -22,7 +22,7 @@ uint16_t ProcessorBase::get_value_of_register(Register r) {
 
 		case Register::A:						return a_;
 		case Register::Flags:					return get_flags();
-		case Register::AF:						return (uint16_t)((a_ << 8) | get_flags());
+		case Register::AF:						return static_cast<uint16_t>((a_ << 8) | get_flags());
 		case Register::B:						return bc_.bytes.high;
 		case Register::C:						return bc_.bytes.low;
 		case Register::BC:						return bc_.full;
@@ -59,7 +59,7 @@ uint16_t ProcessorBase::get_value_of_register(Register r) {
 
 		case Register::IFF1:					return iff1_ ? 1 : 0;
 		case Register::IFF2:					return iff2_ ? 1 : 0;
-		case Register::IM:						return (uint16_t)interrupt_mode_;
+		case Register::IM:						return static_cast<uint16_t>(interrupt_mode_);
 
 		case Register::MemPtr:					return memptr_.full;
 
@@ -72,49 +72,49 @@ void ProcessorBase::set_value_of_register(Register r, uint16_t value) {
 		case Register::ProgramCounter:	pc_.full = value;				break;
 		case Register::StackPointer:	sp_.full = value;				break;
 
-		case Register::A:				a_ = (uint8_t)value;			break;
-		case Register::AF:				a_ = (uint8_t)(value >> 8);		// deliberate fallthrough...
-		case Register::Flags:			set_flags((uint8_t)value);		break;
+		case Register::A:				a_ = static_cast<uint8_t>(value);			break;
+		case Register::AF:				a_ = static_cast<uint8_t>(value >> 8);		// deliberate fallthrough...
+		case Register::Flags:			set_flags(static_cast<uint8_t>(value));		break;
 
-		case Register::B:				bc_.bytes.high = (uint8_t)value;		break;
-		case Register::C:				bc_.bytes.low = (uint8_t)value;			break;
-		case Register::BC:				bc_.full = value;						break;
-		case Register::D:				de_.bytes.high = (uint8_t)value;		break;
-		case Register::E:				de_.bytes.low = (uint8_t)value;			break;
-		case Register::DE:				de_.full = value;						break;
-		case Register::H:				hl_.bytes.high = (uint8_t)value;		break;
-		case Register::L:				hl_.bytes.low = (uint8_t)value;			break;
-		case Register::HL:				hl_.full = value;						break;
+		case Register::B:				bc_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::C:				bc_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::BC:				bc_.full = value;								break;
+		case Register::D:				de_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::E:				de_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::DE:				de_.full = value;								break;
+		case Register::H:				hl_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::L:				hl_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::HL:				hl_.full = value;								break;
 
-		case Register::ADash:			afDash_.bytes.high = (uint8_t)value;	break;
-		case Register::FlagsDash:		afDash_.bytes.low = (uint8_t)value;		break;
-		case Register::AFDash:			afDash_.full = value;					break;
-		case Register::BDash:			bcDash_.bytes.high = (uint8_t)value;	break;
-		case Register::CDash:			bcDash_.bytes.low = (uint8_t)value;		break;
-		case Register::BCDash:			bcDash_.full = value;					break;
-		case Register::DDash:			deDash_.bytes.high = (uint8_t)value;	break;
-		case Register::EDash:			deDash_.bytes.low = (uint8_t)value;		break;
-		case Register::DEDash:			deDash_.full = value;					break;
-		case Register::HDash:			hlDash_.bytes.high = (uint8_t)value;	break;
-		case Register::LDash:			hlDash_.bytes.low = (uint8_t)value;		break;
-		case Register::HLDash:			hlDash_.full = value;					break;
+		case Register::ADash:			afDash_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::FlagsDash:		afDash_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::AFDash:			afDash_.full = value;								break;
+		case Register::BDash:			bcDash_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::CDash:			bcDash_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::BCDash:			bcDash_.full = value;								break;
+		case Register::DDash:			deDash_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::EDash:			deDash_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::DEDash:			deDash_.full = value;								break;
+		case Register::HDash:			hlDash_.bytes.high = static_cast<uint8_t>(value);	break;
+		case Register::LDash:			hlDash_.bytes.low = static_cast<uint8_t>(value);	break;
+		case Register::HLDash:			hlDash_.full = value;								break;
 
-		case Register::IXh:				ix_.bytes.high = (uint8_t)value;		break;
-		case Register::IXl:				ix_.bytes.low = (uint8_t)value;			break;
-		case Register::IX:				ix_.full = value;						break;
-		case Register::IYh:				iy_.bytes.high = (uint8_t)value;		break;
-		case Register::IYl:				iy_.bytes.low = (uint8_t)value;			break;
-		case Register::IY:				iy_.full = value;						break;
+		case Register::IXh:				ix_.bytes.high = static_cast<uint8_t>(value);		break;
+		case Register::IXl:				ix_.bytes.low = static_cast<uint8_t>(value);		break;
+		case Register::IX:				ix_.full = value;									break;
+		case Register::IYh:				iy_.bytes.high = static_cast<uint8_t>(value);		break;
+		case Register::IYl:				iy_.bytes.low = static_cast<uint8_t>(value);		break;
+		case Register::IY:				iy_.full = value;									break;
 
-		case Register::R:				ir_.bytes.low = (uint8_t)value;			break;
-		case Register::I:				ir_.bytes.high = (uint8_t)value;		break;
-		case Register::Refresh:			ir_.full = (uint16_t)value;				break;
+		case Register::R:				ir_.bytes.low = static_cast<uint8_t>(value);		break;
+		case Register::I:				ir_.bytes.high = static_cast<uint8_t>(value);		break;
+		case Register::Refresh:			ir_.full = value;									break;
 
-		case Register::IFF1:			iff1_ = !!value;						break;
-		case Register::IFF2:			iff2_ = !!value;						break;
-		case Register::IM:				interrupt_mode_ = value % 3;			break;
+		case Register::IFF1:			iff1_ = !!value;				break;
+		case Register::IFF2:			iff2_ = !!value;				break;
+		case Register::IM:				interrupt_mode_ = value % 3;	break;
 
-		case Register::MemPtr:			memptr_.full = value;					break;
+		case Register::MemPtr:			memptr_.full = value;			break;
 
 		default: break;
 	}
